Update EHL_02_mainprocess.m
Fixed a bug in the initialization of F_old at the beginning of each time step. This however does not affect any simulation that required 2 or more FBNS iterations per time step, because F_old is overwritten at each FBNS iteration and is only used for one of the convergence criteria.